+++ packages/tokens/src/border/radius/borderRadius4Xl.ts
@@ -1,0 +1,1 @@
+export const borderRadius4Xl = '32px';