+++ packages/tokens/src/border/radius/index.ts
@@ -1,3 +1,8 @@
+export { borderRadius2Xl } from './borderRadius2Xl';
+export { borderRadius3Xl } from './borderRadius3Xl';
+export { borderRadius4Xl } from './borderRadius4Xl';
 export { borderRadiusLg } from './borderRadiusLg';
 export { borderRadiusMd } from './borderRadiusMd';
 export { borderRadiusSm } from './borderRadiusSm';
+export { borderRadiusXl } from './borderRadiusXl';
+export { borderRadiusXs } from './borderRadiusXs';